pyproject.toml
[tool.poetry]
name = "ream2"
version = "1.0.0"
description = "An actor architecture for research software"
authors = ["Binh Vu <binh@toan2.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
t2-yada = "^1.0.4"
timer4 = "^1.0.0"
loguru = "^0.6.0"
graph-wrapper = "^1.5.0"
filelock = "^3.8.0"
orjson = "^3.8.0"
python-slugify = "^6.1.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
